+++ app/src/main/res/values/strings.xml
@@ -20,4 +20,5 @@
     <string name="reminder_removed_success">Geofence removed!</string>
 
     <string name="marker_content_description">Marker</string>
+    <string name="dialog_rationale_permission">We need notification permission to let you know through local push notification when you enter geo fence zone to send</string>
 </resources>
